# src/main/resources/openapi/StudentApiSpec.yaml
openapi: 3.0.0
info:
  title: Student Management API
  version: 1.0.0
  description: API for managing students in the system.
servers:
  - url: http://localhost:8080/api
paths:
  /students:
    post:
      summary: Add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: string
      responses:
        '201':
          description: Student added successfully
        '500':
          description: Failed to add student

    get:
      summary: Get all students
      responses:
        '200':
          description: A list of students
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    studentId:
                      type: string
                    firstName:
                      type: string
                    lastName:
                      type: string
                    email:
                      type: string
                    phone:
                      type: string
                    address:
                      type: string
        '500':
          description: Failed to retrieve students

  /students/{studentId}:
    get:
      summary: Get a student by ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                type: object
                properties:
                  studentId:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                  address:
                    type: string
        '404':
          description: Student not found
        '500':
          description: Failed to retrieve student

    put:
      summary: Update a student by ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                address:
                  type: string
      responses:
        '200':
          description: Student updated successfully
        '500':
          description: Failed to update student

    delete:
      summary: Delete a student by ID
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Student deleted successfully
        '500':
          description: Failed to delete student